LayoutTests/imported/w3c/web-platform-tests/speculation-rules/prefetch/referrer-policy-not-accepted.https_3-last-expected.txt


FAIL with "unsafe-url" referrer policy on cross-site referring page with document rule assert_equals: must not be prefetched expected (undefined) undefined but got (string) "prefetch"
